Rename `io.spine.chords.protobuf` package to `io.spine.chords.proto`
