+++ proto/src/main/kotlin/io/spine/chords/proto/form/InputFieldExts.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.chords.protobuf.form
+package io.spine.chords.proto.form
 
 import com.google.protobuf.Message
 import io.spine.chords.InputField
