+++ proto/src/main/kotlin/io/spine/chords/proto/net/InternetDomainField.kt
@@ -24,30 +24,30 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.chords.protobuf.time
+package io.spine.chords.proto.net
 
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import com.google.protobuf.Timestamp
-import java.time.ZoneId
-import java.time.format.DateTimeFormatter
+import io.spine.chords.ComponentCompanion
+import io.spine.chords.InputField
+import io.spine.chords.exceptionBasedParser
+import io.spine.net.InternetDomain
+import io.spine.net.InternetDomains
 
 /**
- * A text that shows a given date being interpreted in the system time zone.
- *
- * @param date
- *         a date that should be displayed as text.
- * @param pattern
- *         a pattern for formatting the date (see [DateTimeFormatter] for the
- *         formatting syntax).
+ * A field that allows entering an [InternetDomain] value.
  */
-@Composable
-public fun DateText(date: Timestamp, pattern: String = "yyyy-MM-dd") {
-    val instant = date.toInstant()
-    val zoneId = ZoneId.systemDefault()
-    val dateTimeFormatter = DateTimeFormatter.ofPattern(pattern)
+public class InternetDomainField : InputField<InternetDomain>() {
 
-    val zonedDateTime = instant.atZone(zoneId)
-    val dateText = dateTimeFormatter.format(zonedDateTime)
-    Text(dateText)
+    /**
+     * A component instance declaration API.
+     */
+    public companion object : ComponentCompanion<InternetDomainField>({ InternetDomainField() })
+
+    override fun parseValue(rawText: String): InternetDomain = exceptionBasedParser(
+        IllegalArgumentException::class,
+        "Invalid domain syntax"
+    ) {
+        InternetDomains.valueOf(rawText)
+    }
+
+    override fun formatValue(value: InternetDomain): String = value.value
 }
